requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-rosetta==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.35
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
model-bakery==1.17.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
polib==1.2.0
pre-commit==3.6.2
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.1
pytest-django==4.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
